pyproject.toml
[tool.poetry]
name = "magic"
version = "0.1.0"
description = "Major Atmospheric Gamma Imaging Cherenkov Telescope"
authors = ["Ivan <ivanyuminov747@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
torch = "2.2.2"
numpy = "1.26.3"
scikit-learn = "1.3.2"
pyyaml = "6.0.1"
pre-commit = "^4.0.1"
boto3 = "1.34.117"
hydra-core = "1.3.2"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
